md2video/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
pathlib>=1.0.1
PyYAML>=6.0.1
requests>=2.31.0
Pillow>=10.0.0
moviepy>=1.0.3
markdown2>=2.4.10
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver_manager>=4.0.0
markdown>=3.4.0
pydub>=0.25.1